backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# OpenAI compatible client for Grok
openai==1.12.0

# Google APIs
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
